+++ mdast-util-heading-style.js
@@ -34,11 +34,11 @@ function consolidate(depth, relative) {
  * @example
  *   style(); // null
  *
- *   style(mdast.parse('# foo').children[0]); // 'atx'
+ *   style(remark.parse('# foo').children[0]); // 'atx'
  *
- *   style(mdast.parse('# foo #').children[0]); // 'atx-closed'
+ *   style(remark.parse('# foo #').children[0]); // 'atx-closed'
  *
- *   style(mdast.parse('foo\n===').children[0]); // 'setext'
+ *   style(remark.parse('foo\n===').children[0]); // 'setext'
  *
  * @param {Node} node - Node to check.
  * @param {string?} relative - Heading type which we'd wish