+++ history.md
@@ -1,11 +1,11 @@
-<!--mdast setext-->
+<!--remark setext-->
 
 <!--lint disable no-multiple-toplevel-headings-->
 
 1.0.0 / 2015-08-17
 ==================
 
-*   Update dependencies, dev-dependencies ([29e03c7](https://github.com/wooorm/mdast-util-heading-style/commit/29e03c7))
+*   Update dependencies, dev-dependencies ([`29e03c7`](https://github.com/wooorm/mdast-util-heading-style/commit/29e03c7))
 
 0.1.0 / 2015-07-04
 ==================